`GridViewClientScript` extracted
